// src/config.d.ts
export declare const CONFIG: {
    canvas: {
        width: number;
        height: number;
    };
    color: {
        font: string;
        handle: string;
        gradient: {
            color: string;
            stopAt: string;
        }[];
    };
    line: {
        font: string;
        handle: string;
    };
};
//# sourceMappingURL=config.d.ts.map